Move ZSeamConfig helper struct to separate file
As per our code standards, we'd like to have each public class and struct in a separate file. In this case that would've really helped us, since I want to be able to use that struct in a different class too...

Contributes to issue CURA-7928.
