+++ CMakeLists.txt
@@ -196,6 +196,7 @@ set(engine_SRCS # Except main.cpp.
     src/settings/FlowTempGraph.cpp
     src/settings/PathConfigStorage.cpp
     src/settings/Settings.cpp
+    src/settings/ZSeamConfig.cpp
 
     src/utils/AABB.cpp
     src/utils/AABB3D.cpp
